New Heatmap System
Need to rebuild the project around this new heatmap system.

By using a heatmap system we can find all possible places that a ship could be placed on any board state. We can add the probabilities together to get a "best move" and then have the bot make that move