Sunk Ship Tracking
When we sink a ship, we can remove it from the heatmap generation.

This means that if we sink a ship certain spots on the board become "known"

For example, if you already have sunk a 2x1 ship, Some areas of the map can no longer have a ship that fits, meaning that we do not need to waste guesses on them.