Added Strategy Mode
This mode has the goal of eleminating NOT JUST the heatmap locations, but also attempting to reduce the possible ship locations. For example MANY ships are 3 units or larger, so hiting a spot to make a bunch of areas only be able to have 2x1 ships can help speed the game up.